requirements.txt
anyio==4.3.0
asgiref==3.7.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==5.0.3
django-allauth==0.61.1
django-widget-tweaks==1.5.0
fonttools==4.50.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.6
kiwisolver==1.4.5
lxml==4.9.4
matplotlib==3.8.3
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.1
pillow==10.2.0
psycopg-binary==3.1.18
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.2
python-bcb==0.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
reportlab==4.1.0
requests==2.31.0
requests-oauthlib==1.4.0
seaborn==0.13.2
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.6.0
